% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxSimulateModels.R
\name{belong}
\alias{belong}
\title{Comprueba si un objeto pertenece a un paquete concreto}
\usage{
belong(obj, packagename)
}
\arguments{
\item{obj}{Objeto a inspeccionar}

\item{packagename}{Nombre del paquete}
}
\value{
\code{TRUE} si \strong{obj} proviene de \emph{packagename}, \code{FALSE} en otro caso
}
\description{
Comprueba si un objeto pertenece a un paquete concreto
}
\keyword{internal}
