% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworModels.R
\name{Ciclic}
\alias{Ciclic}
\title{Generate a cyclic routing matrix.}
\usage{
Ciclic(n)
}
\arguments{
\item{n}{Integer. Number of nodes.}
}
\description{
Generate a cyclic routing matrix.
}
\keyword{internal}
