% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworModels.R
\name{CLOSED_JACKSON}
\alias{CLOSED_JACKSON}
\title{Functional constructor: Closed Jackson network.}
\usage{
CLOSED_JACKSON(mu, s, p, n)
}
\arguments{
\item{mu}{Numeric vector of service rates.}

\item{s}{Integer vector with the number of servers per node.}

\item{p}{Routing matrix. Rows must sum to 1.}

\item{n}{Integer. Total number of customers in the network.}
}
\value{
A \code{ClosedJackson} object.
}
\description{
Functional constructor: Closed Jackson network.
}
