% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxSimulateModels.R
\name{combineSimulations}
\alias{combineSimulations}
\title{Agrega un listado de simulaciones independientes}
\usage{
combineSimulations(listsims)
}
\arguments{
\item{listsims}{\code{list} con objetos de clase (S3 o R6) simulada}
}
\value{
Un único objeto de la misma clase que \code{listsims[[1]]},
pero con los campos \verb{$out} reemplazados por listas \code{mean/sd/summary}
}
\description{
Combina varias réplicas de un modelo de colas calculando promedio,
desviación típica y un resumen rápido para cada métrica de interés.
}
