% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionAnalysis.R
\name{fitData}
\alias{fitData}
\title{Ajusta varias distribuciones a un vector numérico}
\usage{
fitData(
  data,
  ldistr = c("exp", "norm", "weibull", "unif", "lnorm", "gamma", "beta")
)
}
\arguments{
\item{data}{Vector numérico con las observaciones}

\item{ldistr}{\code{character} con nombres abreviados de distribuciones
(ej. \code{"exp"}, \code{"norm"}, \code{"weibull"}, …)}
}
\value{
Lista de objetos \code{fitdist}.  Clase extra: \code{"FitList"}
}
\description{
Ajusta varias distribuciones a un vector numérico
}
\seealso{
Other DistributionAnalysis: 
\code{\link{goodnessFit}()},
\code{\link{summaryFit}()}
}
\concept{DistributionAnalysis}
