% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovianModel.R
\encoding{UTF-8}
\name{FW}
\alias{FW}
\title{CDF of the time in system \eqn{F_W(x)}}
\usage{
FW(qm, x)
}
\arguments{
\item{qm}{An object that inherits from \code{MarkovianModel}.}

\item{x}{Non‑negative numeric vector.}
}
\description{
CDF of the time in system \eqn{F_W(x)}
}
