% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovianModel.R
\encoding{UTF-8}
\name{FWq}
\alias{FWq}
\title{CDF of the waiting time in queue \eqn{F_{W_q}(x)}}
\usage{
FWq(qm, x)
}
\arguments{
\item{qm}{An object that inherits from \code{MarkovianModel}.}

\item{x}{Non‑negative numeric vector.}
}
\description{
CDF of the waiting time in queue \eqn{F_{W_q}(x)}
}
