% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionAnalysis.R
\name{goodnessFit}
\alias{goodnessFit}
\title{Pruebas chi-cuadrado y KS para cada ajuste}
\usage{
goodnessFit(lfitdata)
}
\arguments{
\item{lfitdata}{Lista producida por \code{\link{fitData}}}
}
\value{
\code{data.frame} con: distribución, estadísticos y \emph{p-values}
}
\description{
Pruebas chi-cuadrado y KS para cada ajuste
}
\seealso{
Other DistributionAnalysis: 
\code{\link{fitData}()},
\code{\link{summaryFit}()}
}
\concept{DistributionAnalysis}
