% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleModels.R
\name{M_M_1}
\alias{M_M_1}
\title{Functional constructor for MM1}
\usage{
M_M_1(lambda = 3, mu = 6)
}
\arguments{
\item{lambda}{Arrival rate (> 0).}

\item{mu}{Service rate (> 0).}
}
\description{
Functional constructor for MM1
}
