% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexModels.R
\name{M_M_1_INF_H}
\alias{M_M_1_INF_H}
\title{Functional constructor for MM1InfH}
\usage{
M_M_1_INF_H(lambda = 0.5, mu = 12, h = 5L)
}
\arguments{
\item{lambda}{Mean arrival rate (> 0).}

\item{mu}{Mean service rate  (> 0).}

\item{h}{Population size (integer >= 1).}
}
\description{
Functional constructor for MM1InfH
}
