% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleModels.R
\name{M_M_1_K}
\alias{M_M_1_K}
\title{Functional constructor for MM1K}
\usage{
M_M_1_K(lambda = 3, mu = 6, k = 2L)
}
\arguments{
\item{lambda}{Arrival rate (> 0).}

\item{mu}{Service rate (> 0).}

\item{k}{Buffer size K (integer >= 0).}
}
\description{
Functional constructor for MM1K
}
