% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexModels.R
\name{M_M_INF}
\alias{M_M_INF}
\title{Functional constructor for MMInf}
\usage{
M_M_INF(lambda = 3, mu = 6)
}
\arguments{
\item{lambda}{Mean arrival rate (> 0).}

\item{mu}{Mean service rate  (> 0).}
}
\description{
Functional constructor for MMInf
}
