% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpleModels.R
\name{M_M_S}
\alias{M_M_S}
\title{Functional constructor for MMS}
\usage{
M_M_S(lambda = 3, mu = 6, s = 2L)
}
\arguments{
\item{lambda}{Arrival rate (> 0).}

\item{mu}{Service rate (> 0).}

\item{s}{Number of servers (integer >= 1).}
}
\description{
Functional constructor for MMS
}
