% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexModels.R
\name{M_M_S_INF_H}
\alias{M_M_S_INF_H}
\title{Functional constructor for MMSInfH}
\usage{
M_M_S_INF_H(lambda = 0.5, mu = 12, s = 2L, h = 5L)
}
\arguments{
\item{lambda}{Mean arrival rate (> 0).}

\item{mu}{Mean service rate  (> 0).}

\item{s}{Servers (integer >= 1).}

\item{h}{Population size (>= s).}
}
\description{
Functional constructor for MMSInfH
}
