% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexModels.R
\name{M_M_S_INF_H_Y}
\alias{M_M_S_INF_H_Y}
\title{Functional constructor for MMSInfHY}
\usage{
M_M_S_INF_H_Y(lambda = 3, mu = 6, s = 3L, h = 5L, y = 3L)
}
\arguments{
\item{lambda}{Mean arrival rate (> 0).}

\item{mu}{Mean service rate  (> 0).}

\item{s}{Servers (>= 1).}

\item{h}{Population size (> 0).}

\item{y}{Number of replacements (>= 1).}
}
\description{
Functional constructor for MMSInfHY
}
