% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovianModel.R
\docType{class}
\encoding{UTF-8}
\name{MarkovianModel}
\alias{MarkovianModel}
\title{\code{MarkovianModel} (abstract R6 class)}
\format{
An \code{R6ClassGenerator} object.
}
\arguments{
\item{arrivalDistribution}{Object created with
\code{distr::Exp(rate)}.}

\item{serviceDistribution}{Object created with
\code{distr::Exp(rate)}.}

\item{n, x, ...}{Arguments passed on to the corresponding methods (see
details above).}
}
\description{
Abstract R6 base class for queueing systems whose
inter-arrival \emph{and} service times follow exponential (Markovian)
distributions.
}
\section{Public fields}{

\describe{
\item{\code{arrivalDistribution}}{An object of class
\code{distr::Exp} that represents inter-arrival times.}
\item{\code{serviceDistribution}}{An object of class
\code{distr::Exp} that represents service times.}
\item{\code{servers}}{\code{integer}. Number of parallel servers.}
\item{\code{out}}{A \code{list} of performance metrics (\eqn{\rho},
\eqn{L}, \eqn{W}, etc.) produced by subclasses.}
}
}

\section{Public methods}{

\describe{
\item{\code{$initialize()}}{Constructor.}
\item{\code{$lambda()} / \code{$mu()}}{Return arrival and service
rates.}
\item{\code{$Pn(n)}}{Steady‑state probability
\eqn{\Pr\{N = n\}}.}
\item{\code{$FW(x)}}{CDF of the time \emph{in the system},
\eqn{F_W(x)}.}
\item{\code{$FWq(x)}}{CDF of the \emph{waiting} time in queue,
\eqn{F_{W_q}(x)}.}
\item{\code{$Qn(n)}}{Probability that the queue length equals
\eqn{n}.}
\item{\code{$maxCustomers()}}{Practical upper bound for the number
of customers the model can hold (may be \eqn{\infty}).}
\item{\code{$print()}}{Pretty printer for the console.}
}
}

\examples{
m <- MM1$new(4, 6)
m$Pn(0:3)
FWq(m, 1)

}
\seealso{
\code{\link{MM1}}, \code{\link{Pn}}, \code{\link{FW}}
}
