% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovianModel.R
\encoding{UTF-8}
\name{maxCustomers}
\alias{maxCustomers}
\title{Upper bound on the number of customers supported}
\usage{
maxCustomers(qm)
}
\arguments{
\item{qm}{An object that inherits from \code{MarkovianModel}.}
}
\description{
Upper bound on the number of customers supported
}
