% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexModels.R
\name{MM1InfH}
\alias{MM1InfH}
\title{MM1InfH-class: M/M/1/Inf/H finite-population queue}
\format{
An \\code{R6Class} generator.
}
\arguments{
\item{lambda}{Mean arrival rate (> 0).}

\item{mu}{Mean service rate  (> 0).}

\item{h}{Population size (integer >= 1).}
}
\description{
R6 class for a single-server Markovian queue with a finite
population of size \\eqn{H}.  Kendall notation: \\emph{M/M/1/Inf/H}.
}
\section{Public methods}{

\describe{
\item{\code{$initialize()}}{Constructor.}
\item{\code{$lambda()} / \code{$mu()}}{Return arrival and service
rates.}
\item{\code{$Pn(n)}}{Steady‑state probability
\eqn{\Pr\{N = n\}}.}
\item{\code{$FW(x)}}{CDF of the time \emph{in the system},
\eqn{F_W(x)}.}
\item{\code{$FWq(x)}}{CDF of the \emph{waiting} time in queue,
\eqn{F_{W_q}(x)}.}
\item{\code{$Qn(n)}}{Probability that the queue length equals
\eqn{n}.}
\item{\code{$maxCustomers()}}{Practical upper bound for the number
of customers the model can hold (may be \eqn{\infty}).}
\item{\code{$print()}}{Pretty printer for the console.}
}
}

\examples{
mm <- MM1InfH$new(0.5, 12, 5)
mm$out$l; Pn(mm, 0:3); FWq(mm, 1)
}
