% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworModels.R
\name{OPEN_JACKSON}
\alias{OPEN_JACKSON}
\title{Functional constructor: Open Jackson network.}
\usage{
OPEN_JACKSON(lambda, mu, s, p)
}
\arguments{
\item{lambda}{Numeric vector of external arrival rates.}

\item{mu}{Numeric vector of service rates.}

\item{s}{Integer vector with the number of servers per node.}

\item{p}{Routing matrix (square, rows sum <= 1).}
}
\value{
An \code{OpenJackson} object.
}
\description{
Functional constructor: Open Jackson network.
}
