% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxSimulateModels.R
\name{ParallelizeSimulations}
\alias{ParallelizeSimulations}
\title{Ejecuta en paralelo \code{nsim} simulaciones de un modelo}
\usage{
ParallelizeSimulations(modelfunction, parameters, nsim = 1L, nproc = 1L)
}
\arguments{
\item{modelfunction}{Función que genera \strong{una} réplica (debe retornar
un objeto con campo \verb{$out})}

\item{parameters}{\code{list} con los argumentos que recibe \code{modelfunction}}

\item{nsim}{Número de réplicas a lanzar}

\item{nproc}{Núcleos a utilizar (>= 1).  Si vale 1 → secuencial}
}
\value{
Una lista con las réplicas, o la única réplica si \code{nsim == 1}
}
\description{
Ejecuta en paralelo \code{nsim} simulaciones de un modelo
}
