% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphicsAux.R
\name{plot_history}
\alias{plot_history}
\title{Quick historic plot for any simulated queue or network}
\usage{
plot_history(x, var = "L", ...)
}
\arguments{
\item{x}{A single simulation object (\code{GG1Sim}, \code{OpenNetSim}, …) \strong{or}
a list of such objects.}

\item{var}{One of \code{"L"}, \code{"Lq"}, \code{"W"}, \code{"Wq"},
\code{"Clients"}, \code{"Intensity"}.}

\item{...}{Extra parameters forwarded to the internal helpers
(e.g., \code{minrange}, \code{maxrange}, \code{depth},
\code{showMean}, \code{showValues}).}
}
\value{
A \pkg{ggplot2} object.
}
\description{
Quick historic plot for any simulated queue or network
}
