% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovianModel.R
\encoding{UTF-8}
\name{Pn}
\alias{Pn}
\title{Steady-state probability \eqn{\Pr\{N = n\}}}
\usage{
Pn(qm, n)
}
\arguments{
\item{qm}{An object that inherits from \code{MarkovianModel}.}

\item{n}{Non‑negative integer vector.}
}
\description{
S3 frontend that delegates to \code{qm$Pn(n)}.
}
