% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovianModel.R
\encoding{UTF-8}
\name{Qn}
\alias{Qn}
\title{Probability that the queue length equals \eqn{n}}
\usage{
Qn(qm, n)
}
\arguments{
\item{qm}{An object that inherits from \code{MarkovianModel}.}

\item{n}{Non‑negative integer vector.}
}
\description{
Probability that the queue length equals \eqn{n}
}
