% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistributionAnalysis.R
\name{summaryFit}
\alias{summaryFit}
\title{Resumen gráfico (densidad, CDF y Q-Q plot)}
\usage{
summaryFit(
  lfitdata,
  graphics = c("ggplot2", "graphics"),
  show = c("all", "dens", "cdf", "qq")
)
}
\arguments{
\item{lfitdata}{Salida de \code{\link{fitData}}}

\item{graphics}{\code{"graphics"} o \code{"ggplot2"}}

\item{show}{\code{"all"}, \code{"dens"}, \code{"cdf"} o \code{"qq"}}
}
\description{
Resumen gráfico (densidad, CDF y Q-Q plot)
}
\seealso{
Other DistributionAnalysis: 
\code{\link{fitData}()},
\code{\link{goodnessFit}()}
}
\concept{DistributionAnalysis}
