% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetworModels.R
\name{Tamdem}
\alias{Tamdem}
\title{Generate a serial (tandem) routing matrix.}
\usage{
Tamdem(n)
}
\arguments{
\item{n}{Integer. Number of nodes.}
}
\description{
Generate a serial (tandem) routing matrix.
}
\keyword{internal}
